# middleware-platform/openapi.yaml
openapi: 3.0.3
info:
  title: Agentic Commerce Platform API
  description: |
    Middleware platform that enables AI agents (ChatGPT, Google, Voice) to process payments.
    Acts as "Plaid for AI Agents" - connecting agent platforms to payment processing.
    
    **Naming migration (safe aliases):**
    - Preferred terminology: `provider` and `prescription`
    - Backward compatible terminology: `merchant` and `product`
    - Endpoints currently accept both field styles during migration.
    
    **Key Features:**
    - Multi-platform support (ACP, AP2, Voice)
    - Multiple payment methods (Stripe, Circle USDC, Mastercard, Visa)
    - Fraud detection and security
    - Healthcare integration (FHIR, insurance, EHR)
  version: 3.0.0
  contact:
    name: API Support
    email: support@doclittle.health
  license:
    name: Proprietary
servers:
  - url: https://web-production-a783d.up.railway.app
    description: Production server
  - url: http://localhost:4000
    description: Local development server
...
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (optional for public endpoints)
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /acp/products:
    get:
      tags: [ACP]
      summary: Get product feed for ChatGPT
      security: []
      parameters:
        - name: merchant_id
          in: query
          required: true
          schema:
            type: string
        - name: query
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Product feed
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      type: object
...
  /ap2/products:
    get:
      tags: [AP2]
      summary: Search products for Google Agent
      security: []
      parameters:
        - name: merchant_id
          in: query
          required: true
          schema:
            type: string
        - name: query
          in: query
          schema:
            type: string
        - name: intent_mandate
          in: query
          schema:
            type: string
            description: JSON-encoded intent mandate
      responses:
        '200':
          description: Products found
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      type: object
...